requirements.txt
flask
mysql-python
flask-sqlalchemy
flask-script
flask-wtf
flask-migrate
py-bcrypt
python-slugify
flask-markdown
flask-uploads


